+++ src/main/java/carRace/controller/RacingController.java
@@ -1,4 +1,8 @@
-package carRace;
+package carRace.controller;
+
+import carRace.domain.Cars;
+import carRace.domain.RacingGame;
+import carRace.view.RacingGameUI;
 
 public class RacingController {
     public static void main(String[] args) {
