+++ src/main/java/carRace/domain/Name.java
@@ -1,4 +1,4 @@
-package carRace;
+package carRace.domain;
 
 public class Name {
     private static final int MAX_NAME_LENGTH = 5;
