+++ src/main/java/carRace/strategy/MovingStrategy.java
@@ -1,4 +1,4 @@
-package carRace;
+package carRace.strategy;
 
 public interface MovingStrategy {
     public boolean movable();
