+++ src/test/java/carRace/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package carRace;
+package carRace.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
