+++ src/test/java/carRace/domain/NameTest.java
@@ -1,4 +1,4 @@
-package carRace;
+package carRace.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
