+++ src/test/java/carRace/strategy/RandomStrategyTest.java
@@ -1,4 +1,4 @@
-package carRace;
+package carRace.strategy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
